pyproject.toml
[tool.poetry]
name = "rewards_agent"
version = "0.1.0"
description = ""
authors = ["Gowtham Satti <sattigowtham1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
python-dotenv = "^1.0.1"
langchain = "^0.3.21"
pymupdf = "^1.25.4"
filetype = "^1.2.0"
python-docx = "^1.1.2"
langchain-experimental = "^0.3.4"
langchain-openai = "^0.3.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
